chore(conformance): extract shared get_process_rss helper
The Linux + macOS platform-specific RSS readers in `batch_pool.rs` and
`server_pool.rs` were byte-identical (28 lines each). Pull them into a
single `crates/conformance/src/process_rss.rs` module and import the
helper in both pools.

The single existing test `get_process_rss_reports_current_process_memory_usage`
moves to the new module. A second Linux-only smoke test verifies the
"`/proc/{pid}/statm` does not exist" path returns `None`.

Behavior-preserving DRY refactor.
