+++ crates/conformance/src/lib.rs
@@ -4,6 +4,7 @@ pub(crate) mod batch_pool;
 pub mod cache;
 pub mod cli;
 pub(crate) mod options_convert;
+pub(crate) mod process_rss;
 pub mod runner;
 pub(crate) mod server_pool;
 pub mod test_filter;
